refactor: move useAuth hook to a separate file for improved organization and maintainability
