+++ src/App.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { BrowserRouter, Routes, Route, Navigate } from 'react-router-dom';
 import { FileText } from 'lucide-react';
-import { useAuth } from './context/AuthContext';
+import { useAuth } from './hooks/api/useAuth';
 import { Sidebar } from './components/layout/Sidebar';
 import { Dashboard } from './pages/Dashboard';
 import { PeoplePage } from './pages/PeoplePage';
