+++ src/hooks/api/useAuth.ts
@@ -1,0 +1,4 @@
+import { useContext } from 'react';
+import { AuthContext } from '@/constants/context';
+
+export const useAuth = () => useContext(AuthContext);
